Updated SignUpButton and SignUpForm components
- Updated SignUpButton component to include a modal for displaying a small box on the screen.
- Modified SignUpForm component to use the UseFormState hook for managing form state.
- Made necessary changes to ensure the modal functionality works for both SignUp and SignUpWithGoogle.
- The modal design will be a small box displayed on the screen.

Note: The UseFormState.js file was added to the hooks folder to handle form state management.
